+++ packages/ui/src/search/sources/gbif.org/index.ts
@@ -1,0 +1,2 @@
+export * from "./GBIFAutocomplete"
+export * from "./GBIFResolve"