+++ apps/api/CHANGELOG.md
@@ -24,6 +24,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security
 
+## [2.4.1] - 2023-04-30
+
+### Fixed
+
+-   `getResolveObjects` was not forwarding when `build` was omitted.
+
 ## [2.4.0] - 2023-04-25
 
 ### Added