+++ apps/contribute/CHANGELOG.md
@@ -14,7 +14,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 -   Made upload error reporting more user-friendly.
--   Upgraded `next` to `13.3.1`.
 
 ### Deprecated
 
@@ -23,6 +22,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 ### Security
+
+## [2.2.0] - 2023-05-02
+
+### Added
+
+-   GBIF to search capabilities.
+
+## [2.1.1] - 2023-04-30
+
+### Changed
+
+-   Upgraded `next` to `13.3.1`.
 
 ## [2.1.0] - 2023-04-25