+++ apps/edit/CHANGELOG.md
@@ -13,8 +13,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
--   Upgraded `next` to `13.3.1`.
-
 ### Deprecated
 
 ### Fixed
@@ -22,6 +20,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 ### Security
+
+## [1.2.0] - 2023-05-02
+
+### Added
+
+-   GBIF to search capabilities.
+
+## [1.1.1] - 2023-04-30
+
+### Changed
+
+-   Upgraded `next` to `13.3.1`.
 
 ## [1.1.0] - 2023-04-25